[log] Assertions so nice you'll want some to fail
I've noticed I often get:

[S 0] Failed assertion on build/opt/zsim.cpp:119

Then go turn the assertion into an assert_msg to get the values, and
leave it there.

Instead, a failing assertion now prints:

[S 0] Failed assertion on build/opt/zsim.cpp:119 'cid >= zinfo->numCores' (with '0 >= 1')

Note this shows the expression AND the captured values. This renders
most assert_msgs unnecessary, and saves you a trip to the code. It
also does not decrease performance.

Capturing expression values works for most typical expressions,
including comparisons, arithmetic and logical expressions. It relies on
truly evil operator overloading. A side-effect of this is that if
components of an expression don't go through a stringsteam (basically
comparisons between iterators, e.g., assert(it != vec.end())), the
expression needs to be parenthesized, i.e., assert((it != vec.end())).
These will be fixed in the next commit.