[slab_alloc] New garbage-collected slab allocator for timing events
The old slab allocator used a heuristic based on simulated cycles to
decide when it was safe to deallocate a slab. This was not robust, as
some events (e.g., PUTs) do not stall cores and can fall behind, and
were causing memory corruption.

Instead, the new slab allocator uses reference counting to find when
all events in the slab have finished. It leverages that slab-allocated
events must call done() after finishing execution. It may cause memory
leaks (e.g., if events are created but not connected to other events),
but should be robust against memory corruption.

This changeset "breaks" the OOO core, as the recorder stores a pointer
to an old event that may have been deallocated.

Conflicts:
	src/phase_slab_alloc.h