[virt] Add system call virtualization for sys_alarm.
This enables using alarm() to properly receive a signal after the requested
amount of virtual time has passed. A new signal queue is added to the
scheduler to support this.

From a performance perspective this change should be minimal, however
it does require checking a queue for emptiness at each phase.

Tested with a simple signal-receiving app and pin 2.14 (71313).